requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
asyncpg==0.30.0
SQLAlchemy==2.0.41
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
python-dotenv==1.1.0
pydantic==2.11.5